move js code to separate file
